simple cart project

+++ Shopping Cart/Models/Cart.swift
@@ -10,5 +10,5 @@ import Foundation
 struct Cart: Identifiable {
     var id = UUID()
     var productId: UUID
-    var quantity: Int
+    var quantity: Int16
 }
