+++ Shopping Cart/Models/Category.swift
@@ -10,5 +10,5 @@ import Foundation
 struct Category: Identifiable {
     var id = UUID()
     var name: String
-    var color: Int
+    var color: Int16
 }
